requirements.txt
streamlit
google-generativeai
python-dotenv
pymongo
dnspython

# Use a more compatible version for Windows
numpy==1.24.3
pandas==2.0.3

# Optional for production
gunicorn==21.2.0

